gcp/cloud_run/requirements.txt
flask>=3.0.0
wandb>=0.17.0
google-cloud-secret-manager>=2.20.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.48.0
google-cloud-scheduler>=2.13.0
kfp>=2.7.0
pyyaml>=6.0
